split ui from core PlayerDesk
